+++ example/src/main.tsx
@@ -1,18 +1,23 @@
 import { TangledContextProvider } from '@tangled3/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
 
+const queryClient = new QueryClient();
+
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <TangledContextProvider
-      config={{
-        projectName: 'Tangled Example',
-        chainConfigs: {},
-      }}
-    >
-      <App />
-    </TangledContextProvider>
+    <QueryClientProvider client={queryClient}>
+      <TangledContextProvider
+        config={{
+          projectName: 'Tangled Example',
+          chainConfigs: {},
+        }}
+      >
+        <App />
+      </TangledContextProvider>
+    </QueryClientProvider>
   </React.StrictMode>,
 );